+++ NEWS.md
@@ -1,2 +1,6 @@
+# ruminate 0.2.0 
+- Added MB (model builder) module with
+  - Support for rxode2 and NONMEM models
+
 # ruminate 0.1.1
 - Initial release